+++ source/java/ch/systemsx/cisd/hdf5/HDF5ByteReader.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2007 - 2014 ETH Zuerich, CISD and SIS.
+ * Copyright 2007 - 2018 ETH Zuerich, CISD and SIS.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,23 +19,22 @@ package ch.systemsx.cisd.hdf5;
 import static ch.systemsx.cisd.hdf5.MatrixUtils.cardinalityBoundIndices;
 import static ch.systemsx.cisd.hdf5.MatrixUtils.checkBoundIndices;
 import static ch.systemsx.cisd.hdf5.MatrixUtils.createFullBlockDimensionsAndOffset;
-import static ch.systemsx.cisd.hdf5.hdf5lib.HDF5Constants.H5T_ARRAY;
-import static ch.systemsx.cisd.hdf5.hdf5lib.HDF5Constants.H5T_NATIVE_INT8;
+import static hdf.hdf5lib.HDF5Constants.H5T_ARRAY;
+import static hdf.hdf5lib.HDF5Constants.H5T_NATIVE_INT8;
 
 import java.util.Arrays;
 import java.util.Iterator;
 
-import ncsa.hdf.hdf5lib.exceptions.HDF5JavaException;
-import ncsa.hdf.hdf5lib.exceptions.HDF5LibraryException;
-import ncsa.hdf.hdf5lib.exceptions.HDF5SpaceRankMismatch;
-
+import hdf.hdf5lib.exceptions.HDF5JavaException;
+import hdf.hdf5lib.exceptions.HDF5LibraryException;
+import ch.ethz.sis.hdf5.exceptions.HDF5SpaceRankMismatch;
 import ch.systemsx.cisd.base.mdarray.MDArray;
 import ch.systemsx.cisd.base.mdarray.MDByteArray;
 import ch.systemsx.cisd.hdf5.HDF5BaseReader.DataSpaceParameters;
 import ch.systemsx.cisd.hdf5.HDF5DataTypeInformation.DataTypeInfoOptions;
 import ch.systemsx.cisd.hdf5.cleanup.ICallableWithCleanUp;
 import ch.systemsx.cisd.hdf5.cleanup.ICleanUpRegistry;
-import ch.systemsx.cisd.hdf5.hdf5lib.HDF5Constants;
+import hdf.hdf5lib.HDF5Constants;
 
 /**
  * The implementation of {@link IHDF5ByteReader}.
@@ -75,9 +74,9 @@ class HDF5ByteReader implements IHDF5ByteReader
                 @Override
                 public Byte call(ICleanUpRegistry registry)
                 {
-                    final int objectId =
+                    final long objectId =
                             baseReader.h5.openObject(baseReader.fileId, objectPath, registry);
-                    final int attributeId =
+                    final long attributeId =
                             baseReader.h5.openAttribute(objectId, attributeName, registry);
                     final byte[] data =
                             baseReader.h5.readAttributeAsByteArray(attributeId, H5T_NATIVE_INT8, 1);
@@ -100,7 +99,7 @@ class HDF5ByteReader implements IHDF5ByteReader
                         @Override
                         public byte[] call(ICleanUpRegistry registry)
                         {
-                            final int objectId =
+                            final long objectId =
                                     baseReader.h5.openObject(baseReader.fileId, objectPath,
                                             registry);
                             return getByteArrayAttribute(objectId, attributeName, registry);
@@ -123,7 +122,7 @@ class HDF5ByteReader implements IHDF5ByteReader
                         @Override
                         public MDByteArray call(ICleanUpRegistry registry)
                         {
-                            final int objectId =
+                            final long objectId =
                                     baseReader.h5.openObject(baseReader.fileId, objectPath,
                                             registry);
                             return getByteMDArrayAttribute(objectId, attributeName, registry);
@@ -160,7 +159,7 @@ class HDF5ByteReader implements IHDF5ByteReader
                 @Override
                 public Byte call(ICleanUpRegistry registry)
                 {
-                    final int dataSetId = 
+                    final long dataSetId = 
                             baseReader.h5.openDataSet(baseReader.fileId, objectPath, registry);
                     final byte[] data = new byte[1];
                     baseReader.h5.readDataSet(dataSetId, H5T_NATIVE_INT8, data);
@@ -181,7 +180,7 @@ class HDF5ByteReader implements IHDF5ByteReader
                 @Override
                 public byte[] call(ICleanUpRegistry registry)
                 {
-                    final int dataSetId = 
+                    final long dataSetId = 
                             baseReader.h5.openDataSet(baseReader.fileId, objectPath, registry);
                     return readByteArray(dataSetId, registry);
                 }
@@ -189,7 +188,7 @@ class HDF5ByteReader implements IHDF5ByteReader
         return baseReader.runner.call(readCallable);
     }
 
-    private byte[] readByteArray(int dataSetId, ICleanUpRegistry registry)
+    private byte[] readByteArray(long dataSetId, ICleanUpRegistry registry)
     {
         try
         {
@@ -205,7 +204,7 @@ class HDF5ByteReader implements IHDF5ByteReader
                     && ex.getMinorErrorNumber() == HDF5Constants.H5E_CANTINIT)
             {
                 // Check whether it is an array data type.
-                final int dataTypeId = baseReader.h5.getDataTypeForDataSet(dataSetId, registry);
+                final long dataTypeId = baseReader.h5.getDataTypeForDataSet(dataSetId, registry);
                 if (baseReader.h5.getClassType(dataTypeId) == HDF5Constants.H5T_ARRAY)
                 {
                     return readByteArrayFromArrayType(dataSetId, dataTypeId, registry);
@@ -215,13 +214,13 @@ class HDF5ByteReader implements IHDF5ByteReader
         }
     }
 
-    private byte[] readByteArrayFromArrayType(int dataSetId, final int dataTypeId,
+    private byte[] readByteArrayFromArrayType(long dataSetId, final long dataTypeId,
             ICleanUpRegistry registry)
     {
-        final int spaceId = baseReader.h5.createScalarDataSpace();
+        final long spaceId = baseReader.h5.createScalarDataSpace();
         final int[] dimensions = baseReader.h5.getArrayDimensions(dataTypeId);
         final byte[] data = new byte[HDF5Utils.getOneDimensionalArraySize(dimensions)];
-        final int memoryDataTypeId =
+        final long memoryDataTypeId =
                 baseReader.h5.createArrayType(H5T_NATIVE_INT8, data.length, registry);
         baseReader.h5.readDataSet(dataSetId, memoryDataTypeId, spaceId, spaceId, data);
         return data;
@@ -239,12 +238,12 @@ class HDF5ByteReader implements IHDF5ByteReader
                 @Override
                 public int[] call(ICleanUpRegistry registry)
                 {
-                    final int dataSetId = 
+                    final long dataSetId = 
                             baseReader.h5.openDataSet(baseReader.fileId, objectPath, registry);
                     final DataSpaceParameters spaceParams =
                             baseReader.getBlockSpaceParameters(dataSetId, memoryOffset, array
                                     .dimensions(), registry);
-                    final int nativeDataTypeId =
+                    final long nativeDataTypeId =
                             baseReader.getNativeDataTypeId(dataSetId, H5T_NATIVE_INT8, registry);
                     baseReader.h5.readDataSet(dataSetId, nativeDataTypeId, 
                             spaceParams.memorySpaceId, spaceParams.dataSpaceId, array.
@@ -268,12 +267,12 @@ class HDF5ByteReader implements IHDF5ByteReader
                 @Override
                 public int[] call(ICleanUpRegistry registry)
                 {
-                    final int dataSetId = 
+                    final long dataSetId = 
                             baseReader.h5.openDataSet(baseReader.fileId, objectPath, registry);
                     final DataSpaceParameters spaceParams =
                             baseReader.getBlockSpaceParameters(dataSetId, memoryOffset, array
                                     .dimensions(), offset, blockDimensions, registry);
-                    final int nativeDataTypeId =
+                    final long nativeDataTypeId =
                             baseReader.getNativeDataTypeId(dataSetId, H5T_NATIVE_INT8, registry);
                     baseReader.h5.readDataSet(dataSetId, nativeDataTypeId, 
                             spaceParams.memorySpaceId, spaceParams.dataSpaceId, array
@@ -309,7 +308,7 @@ class HDF5ByteReader implements IHDF5ByteReader
                 @Override
                 public byte[] call(ICleanUpRegistry registry)
                 {
-                    final int dataSetId = 
+                    final long dataSetId = 
                             baseReader.h5.openDataSet(baseReader.fileId, objectPath, registry);
                     final DataSpaceParameters spaceParams =
                             baseReader.getSpaceParameters(dataSetId, offset, blockSize, registry);
@@ -444,7 +443,7 @@ class HDF5ByteReader implements IHDF5ByteReader
                 @Override
                 public MDByteArray call(ICleanUpRegistry registry)
                 {
-                    final int dataSetId = 
+                    final long dataSetId = 
                             baseReader.h5.openDataSet(baseReader.fileId, objectPath, registry);
                     return readByteMDArray(dataSetId, registry);
                 }
@@ -452,7 +451,7 @@ class HDF5ByteReader implements IHDF5ByteReader
         return baseReader.runner.call(readCallable);
     }
 
-    MDByteArray readByteMDArray(int dataSetId, ICleanUpRegistry registry)
+    MDByteArray readByteMDArray(long dataSetId, ICleanUpRegistry registry)
     {
         try
         {
@@ -468,7 +467,7 @@ class HDF5ByteReader implements IHDF5ByteReader
                     && ex.getMinorErrorNumber() == HDF5Constants.H5E_CANTINIT)
             {
                 // Check whether it is an array data type.
-                final int dataTypeId = baseReader.h5.getDataTypeForDataSet(dataSetId, registry);
+                final long dataTypeId = baseReader.h5.getDataTypeForDataSet(dataSetId, registry);
                 if (baseReader.h5.getClassType(dataTypeId) == HDF5Constants.H5T_ARRAY)
                 {
                     return readByteMDArrayFromArrayType(dataSetId, dataTypeId, registry);
@@ -478,16 +477,16 @@ class HDF5ByteReader implements IHDF5ByteReader
         }
     }
 
-    private MDByteArray readByteMDArrayFromArrayType(int dataSetId, final int dataTypeId,
+    private MDByteArray readByteMDArrayFromArrayType(long dataSetId, final long dataTypeId,
             ICleanUpRegistry registry)
     {
         final int[] arrayDimensions = baseReader.h5.getArrayDimensions(dataTypeId);
-        final int memoryDataTypeId =
+        final long memoryDataTypeId =
                 baseReader.h5.createArrayType(H5T_NATIVE_INT8, arrayDimensions, registry);
         final DataSpaceParameters spaceParams = baseReader.getSpaceParameters(dataSetId, registry);
         if (spaceParams.blockSize == 0)
         {
-            final int spaceId = baseReader.h5.createScalarDataSpace();
+            final long spaceId = baseReader.h5.createScalarDataSpace();
             final byte[] data = new byte[MDArray.getLength(arrayDimensions)];
             baseReader.h5.readDataSet(dataSetId, memoryDataTypeId, spaceId, spaceId, data);
             return new MDByteArray(data, arrayDimensions);
@@ -586,7 +585,7 @@ class HDF5ByteReader implements IHDF5ByteReader
                 @Override
                 public MDByteArray call(ICleanUpRegistry registry)
                 {
-                    final int dataSetId = 
+                    final long dataSetId = 
                             baseReader.h5.openDataSet(baseReader.fileId, objectPath, registry);
                     try
                     {
@@ -618,7 +617,7 @@ class HDF5ByteReader implements IHDF5ByteReader
         return baseReader.runner.call(readCallable);
     }
     
-    private MDByteArray readMDArrayBlockOfArrays(final int dataSetId, final int[] blockDimensions,
+    private MDByteArray readMDArrayBlockOfArrays(final long dataSetId, final int[] blockDimensions,
             final long[] offset, final HDF5DataSetInformation info, final int spaceRank,
             final ICleanUpRegistry registry)
     {
@@ -649,7 +648,7 @@ class HDF5ByteReader implements IHDF5ByteReader
                 baseReader.getSpaceParameters(dataSetId, spaceOfs, spaceBlockDimensions, registry);
         final byte[] dataBlock =
                 new byte[spaceParams.blockSize * info.getTypeInformation().getNumberOfElements()];
-        final int memoryDataTypeId =
+        final long memoryDataTypeId =
                 baseReader.h5.createArrayType(H5T_NATIVE_INT8, info.getTypeInformation()
                         .getDimensions(), registry);
         baseReader.h5.readDataSet(dataSetId, memoryDataTypeId, spaceParams.memorySpaceId,
@@ -748,14 +747,14 @@ class HDF5ByteReader implements IHDF5ByteReader
             };
     }
 
-    byte[] getByteArrayAttribute(final int objectId, final String attributeName,
+    byte[] getByteArrayAttribute(final long objectId, final String attributeName,
             ICleanUpRegistry registry)
     {
-        final int attributeId =
+        final long attributeId =
                 baseReader.h5.openAttribute(objectId, attributeName, registry);
-        final int attributeTypeId =
+        final long attributeTypeId =
                 baseReader.h5.getDataTypeForAttribute(attributeId, registry);
-        final int memoryTypeId;
+        final long memoryTypeId;
         final int len;
         if (baseReader.h5.getClassType(attributeTypeId) == H5T_ARRAY)
         {
@@ -785,16 +784,16 @@ class HDF5ByteReader implements IHDF5ByteReader
         return data;
     }
 
-    MDByteArray getByteMDArrayAttribute(final int objectId,
+    MDByteArray getByteMDArrayAttribute(final long objectId,
             final String attributeName, ICleanUpRegistry registry)
     {
         try
         {
-            final int attributeId =
+            final long attributeId =
                     baseReader.h5.openAttribute(objectId, attributeName, registry);
-            final int attributeTypeId =
+            final long attributeTypeId =
                     baseReader.h5.getDataTypeForAttribute(attributeId, registry);
-            final int memoryTypeId;
+            final long memoryTypeId;
             final int[] arrayDimensions;
             if (baseReader.h5.getClassType(attributeTypeId) == H5T_ARRAY)
             {
